requirements.txt
numpy
matplotlib
pandas
pdfminer.six
tika
nltk
csv
bs4
gensim
sklearn
tqdm 
seaborn
operator
json
itertools
os
matplotlib
tabulate
IPython.display 
